2025.01.24 SRX1431-100(CMT2310A_426.075)_V006 ①修正测试模式中发射才能接收的问题。 ②工作模式426.075/429.350/429.550扫描接收，429.350发射数据

+++ user_src/ram.c
@@ -97,7 +97,7 @@ UINT8 Tx_Rx_mode=0;
                                                0x00000000,0x00000000,0x00000000,0x00000000,
                                                0x00000000,0x00000000,0x00000000,0x00000000,
                                                };*/
-UINT8 Time_APP_blank_TX = 0;
+UINT16 Time_APP_blank_TX = 0;
 UINT8  TIME_10ms=0;
 UINT16 TIME_auto_useful = 0;
 UINT8 FREQ_auto_useful = 0;
